requirements.txt
Babel==2.3.4
-e git+https://github.com/PaulUithol/backbone-tastypie.git@88bd191a0094922cd37a894e6c36fdc0a836b01e#egg=Backbone_tastypie
cmsplugin-contact==1.0.0
cmsplugin-filer==0.9.5
coverage==3.6
diff-match-patch==20121119
Django==1.10
django-annoying==0.7.9
django-bower==5.1.0
django-classy-tags==0.7.2
django-cms==2.4.3
django-contrib-comments==1.7.2
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-durationfield==0.5.2
django-filer==1.2.4
django-fluent-comments==0.9.2
django-grappelli==2.8.1
django-import-export==0.1.5
django-mptt==0.5.2
django-phonenumber-field==1.1.0
django-polymorphic==0.9.2
-e git+git@github.com:jrutila/django-reportengine@4e98c4d746d6fe9849da3869e915e06f214f46cb#egg=django_reportengine
django-reusableapps==0.1.1
django-reversion==2.0.6
-e git+git@github.com:jrutila/django-schedule.git@4fbe8e3074ec09feb2539af967d491cc5a7ec721#egg=django_scheduler
django-sekizai==0.9.0
-e git://github.com/jrutila/django-settings.git@e63e390ed7154bc4f026bf21588b26cbb5159a8b#egg=django_settings
-e git+git@github.com:jrutila/django-shop.git@b5a132f7377e57fc6bafa1bdade81935d643206d#egg=django_shop
-e git+git@github.com:jrutila/django-stables.git@0578f1a9e5099fd5c11fde10809ef6a9cffa0419#egg=django_stables
django-tastypie==0.13.3
django-tenant-schemas==1.4.4
django-twilio==0.8.0
djangorestframework==3.4.1
easy-thumbnails==2.3
html5lib==0.999999999
httplib2==0.9.2
importlib==1.0.3
jsonfield==1.0.3
newrelic==2.68.0.50
phonenumberslite==7.5.1
Pillow==3.3.0
psycopg2==2.6.2
python-dateutil==2.5.3
python-memcached==1.58
python-mimeparse==1.5.2
pytz==2016.6.1
requests==2.10.0
six==1.10.0
solid-i18n==1.3.0
South==1.0
tablib==0.11.2
twilio==5.4.0
Unidecode==0.4.19
vobject==0.9.2
webencodings==0.5
